MoodRecord and MoodRecordDAO
This has the model class for mood records and the DAO to read/write them. Both classes have tests (which all pass). Once merged the app will support reading and writing records.

The Quantifiable interface is there because it's needed for the implementation of my design pattern. It's an interface that specifies the data/methods a model class needs to have in order to be displayed in a chart.